yyt: add error detection for .y and .l, unified error output format

+++ test/error_test/miss_semi.c
@@ -1,6 +1,6 @@
 int main()
 {
-    float a[10];
+    int a[10];
     int i;
     a[3] = 3 if (a[2] == 0)
         i = 1 else i = 0;
